Add WASI support for CLI
